# Specify hints to TSDB querier when querying series